# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API Despesas Cartão",
        "description": "Descrição da API para gerenciar despesas de cartão",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Insere na base de dados os registros do novo usuário.",
                "operationId": "f35f51ea578a0d55fd5a86139ea9d086",
                "responses": {
                    "200": {
                        "description": "Successfully Done!"
                    },
                    "500": {
                        "description": "Unexpected error!"
                    },
                    "422": {
                        "description": "User already registered!"
                    },
                    "423": {
                        "description": "Email already registered with another user!"
                    },
                    "424": {
                        "description": "Request error, validate the data and try again!"
                    }
                }
            }
        },
        "/api/users/{id_user}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Lista os dados somente do usuário selecionado.",
                "operationId": "16e5be37127bdde0f97fbc61b53c01bc",
                "responses": {
                    "420": {
                        "description": "No registration!"
                    },
                    "425": {
                        "description": "Undefined User!"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualiza os dados somente do usuário selecionado já cadastrado.",
                "operationId": "603fa937b43a1c37da647a9a870e22b4",
                "responses": {
                    "200": {
                        "description": "Successfully Done!"
                    },
                    "500": {
                        "description": "Unexpected error!"
                    },
                    "424": {
                        "description": "Request error, validate the data and try again!"
                    },
                    "425": {
                        "description": "Undefined User!"
                    },
                    "426": {
                        "description": "There is already a user with these credentials registered!"
                    },
                    "427": {
                        "description": "There is already a user with this email registered!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Exclui na base de dados os registros do usuário selecionado.",
                "operationId": "5e8170d9db1d778080bab6962b922ccd",
                "responses": {
                    "200": {
                        "description": "Successfully Done!"
                    },
                    "500": {
                        "description": "Unexpected error!"
                    },
                    "424": {
                        "description": "Request error, validate the data and try again!"
                    },
                    "425": {
                        "description": "Undefined User!"
                    }
                }
            }
        }
    },
...
}